pyproject.toml
[tool.poetry]
name = "youtubevideotomp3"
version = "0.1.0"
description = ""
authors = ["Erick Sanabria <martinoli944@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
clipboard = "^0.0.4"
yt-dlp = "^2023.11.16"
python-dotenv = "^1.0.0"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
